hashmap: turn off mempool logic for outside users
This turns off the mempool logic of our hashmap implementation for all
external users, but leaves it on for our own programs.

This is implemented by double compilation of the relevant source files:
when -DUSE_HASHMAP_MEMPOOL=1 is set we'll use the mempool, and if
-DUSE_HASHMAP_MEMPOOL=0 is used we'll not. We then define two meson
build objects `libhashmap` and `libhashmap_nomempool` that set this
option differently. Finally, the build targets that previously imported
libbasic or libshared blanket is now updated to pick one or the other
depending whether the code shows up in external processes or not.
Specifically:

- libsystemd.so, the NSS and PAM modules use the libhashmap_nomempool
- libsystemd-shared-XYZ.so uses libhashmap

The latter is only used by our internal tools, the former to everybody
else, hence we get the desired behaviour.

Follow-up for #9792