udev-builtin-net_id: don't rename the OLPC mesh devices
The Marvell Libertas 88W8388 driver, as present on OLPC XO-1 laptops,
creates two WLAN interfaces: one for regular operation and another for
its special mesh capability.

udev-builtin-net_id currently attempts to give both devices the same
name, causing one of them to always lose the race:

  error changing net interface name 'wlan0' to 'wlx0017c41182ff': File exists
  could not rename interface '7' from 'wlan0' to 'wlx0017c41182ff': File exists

Let's just keep those devices, both of them, alone.

An alternative would be to give the mesh device a prefix different from
"wl". However, the wlan currently gets a rather silly name anyways. The
device, despite sitting on an USB bus, is in fact not removable, but the
platform fails to expose that in sysfs and fixing that would be
non-trivial.