bls: Update introduction wrt. modern firmware
Most open-source firmware doesn't implement BIOS or UEFI interfaces but
includes a boot loader instead. Also BIOS and UEFI are concepts mostly
only present on x86. Thus, the concept of the Boot Loader Specification
can easily be extended to allow modern firmware to directly boot the OS
without an intermediate, on-disk boot loader. This is an area that cur-
rently misses any standard.

The major necessary change to the Boot Loader Specification is to also
cover installation and live media and not only OS installations.

Signed-off-by: Nico Huber <nico.h@gmx.de>