parse: Add CategoricalBool and parser
This change allows a boolean to be "categorical", with the meaning for
that forcing to be determined on a field by field basis. The first user
is likely to be cgroup controller related settings, where this allows
forcibly disabling a particular controller within a specific subtree in
the cgroup hierarchy. See issues #7624 and #10473 for why this is useful
in that case.

Currently this isn't hooked up to any config options, that will happen
after.

Test Plan:

    % ninja -C build test
    % grep -e parse-util -e conf-parser build/meson-logs/testlog.txt | grep OK
    325/493 test-parse-util                         OK       0.02 s
    384/493 test-conf-parser                        OK       1.87 s