random-seed: prepare for allowing crediting entropy from seed file
This reworks random-seed to use the RNDADDENTROPY ioctl. With an entropy_count
of 0, this is exactly equivalent to writing the bytes to /dev/urandom, so this
patch in itself does not change behaviour. However, in subsequent patches we
will store an estimate of the amount of entropy of the random data in an xattr
associated to the file, and teach systemd-random-seed to take an option

  --entropy-credit <ratio>

(and/or environment variable SYSTEMD_ENTROPY_CREDIT) to indicate how much of
that entropy should be credited to the the kernel's random pool.