udev-node: Don't reset atime and mtime when fixing up a node
This basically reverts 5607391481.

When fixing up a node we might do chmod(), chown(), etc. These
change ctime for the node. But changing atime and mtime is
unjustified. We are not really accessing the node nor changing
its contents. Also, for updating ctime() we should rely on
aforementioned syscalls instead of forcing change ourselves.

This bug is particularly visible when not applying any rule but
the node Xtime is changed for no good reason.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>