networkd: remove a weird assertion from network_load_one
The assertion was added in dbffab87f1504abc9f but I'm not exactly sure
why. Now it just crashes a fuzzer I wrote. I, of course, could fix the fuzzer
so that it would generate names ending in ".network" if anyone would tell
me why the assertion should be kept here.