tree-wide: introduce _cleanup_(pager_closep) and use it (almost) everywhere
I left systemd itself with pager_close(), because there is a long sequence of cleanup
events there, and I wasn't entirely sure whether the order matters.

In all other cases, pager_close() is either the only cleanup step, or pager_close()
is only called if the other steps won't matter (e.g. because pager_open() is only
called from help()), or the cleanup steps are independent and reordering them is OK.
In a few cases, the bus needs to be cleaned up before the pager. Using _cleanup_ for
pager_close() still does this.